Dock/TaskBar icon progress - move jobs listener registration to NotifierProvider
